refactor ConfigurationFile namespace
